move implementations from header to cpp file
