+++ app/brewblox/blox/SparkIoBase.h
@@ -19,11 +19,8 @@
 
 #pragma once
 
-#include "ActuatorDigital.h"
-#include "Board.h"
 #include "IoArray.h"
-#include "IoArrayHelpers.h"
-#include "blox/Block.h"
+#include "pinmap_hal.h" // pin_t defined here
 
 class SparkIoBase : public IoArray {
 protected:
@@ -37,55 +34,9 @@ public:
 
     // generic ArrayIO interface
     virtual bool
-    senseChannelImpl(uint8_t channel, State& result) const override final
-    {
-        auto pin = channelToPin(channel);
-        if (pin != static_cast<decltype(pin)>(-1)) {
-            result = pinReadFast(pin) != 0 ? State::Active : State::Inactive;
-            return true;
-        }
-        return false;
-    }
-
+    senseChannelImpl(uint8_t channel, State& result) const override final;
     virtual bool
-    writeChannelImpl(uint8_t channel, ChannelConfig config) override final
-    {
-        auto pin = channelToPin(channel);
-        if (pin != static_cast<decltype(pin)>(-1)) {
-#ifdef PIN_V3_TOP1_DIR
-            if (pin == PIN_V3_TOP1) {
-                bool isOutput = (config == ChannelConfig::ACTIVE_HIGH || config == ChannelConfig::ACTIVE_LOW);
-                HAL_Pin_Mode(PIN_V3_TOP1_DIR, OUTPUT);
-                digitalWriteFast(PIN_V3_TOP1_DIR, isOutput);
-            }
-#endif
-#ifdef PIN_V3_TOP2_DIR
-            if (pin == PIN_V3_TOP2) {
-                bool isOutput = (config == ChannelConfig::ACTIVE_HIGH || config == ChannelConfig::ACTIVE_LOW);
-                HAL_Pin_Mode(PIN_V3_TOP2_DIR, OUTPUT);
-                digitalWriteFast(PIN_V3_TOP2_DIR, isOutput);
-            }
-#endif
-            switch (config) {
-            case ChannelConfig::ACTIVE_HIGH:
-                HAL_Pin_Mode(pin, OUTPUT);
-                digitalWriteFast(pin, true);
-                break;
-            case ChannelConfig::ACTIVE_LOW:
-                HAL_Pin_Mode(pin, OUTPUT);
-                digitalWriteFast(pin, false);
-                break;
-            case ChannelConfig::INPUT:
-            case ChannelConfig::UNUSED:
-            case ChannelConfig::UNKNOWN:
-                HAL_Pin_Mode(pin, INPUT_PULLDOWN);
-                break;
-            }
-            return true;
-        }
-        return false;
-    }
-
+    writeChannelImpl(uint8_t channel, ChannelConfig config) override final;
     virtual bool
     supportsFastIo() const override final
     {
