+++ app/brewblox/blox/TempSensorCombiBlock.h
@@ -1,12 +1,30 @@
+/*
+ * Copyright 2020 BrewPi B.V.
+ *
+ * This file is part of BrewBlox
+ *
+ * BrewBlox is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This library is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with BrewBlox.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
 #pragma once
 
-#include "TempSensorCombi.h"
-
 #include "TempSensor.h"
+#include "TempSensorCombi.h"
 #include "blox/Block.h"
-#include "blox/FieldTags.h"
 #include "cbox/CboxPtr.h"
 #include "proto/cpp/TempSensorCombi.pb.h"
+#include <vector>
 
 class TempSensorCombiBlock : public Block<BrewBloxTypes_BlockType_TempSensorCombi> {
 private:
@@ -22,89 +40,21 @@ public:
 
     virtual ~TempSensorCombiBlock() = default;
 
-    virtual cbox::CboxError
-    streamFrom(cbox::DataIn& in) override final
-    {
-        blox_TempSensorCombi newData = blox_TempSensorCombi_init_zero;
-        cbox::CboxError result = streamProtoFrom(in, &newData, blox_TempSensorCombi_fields, blox_TempSensorCombi_size);
-        if (result == cbox::CboxError::OK) {
-            sensor.func = TempSensorCombi::CombineFunc(newData.combinefunc);
-            inputs.clear();
-            sensor.inputs.clear();
-            inputs.reserve(newData.sensors_count);
-            sensor.inputs.reserve(newData.sensors_count);
-            for (uint8_t i = 0; i < newData.sensors_count && i < 8; i++) {
-                inputs.push_back(cbox::CboxPtr<TempSensor>(objectsRef, newData.sensors[i]));
-            }
-            for (auto& i : inputs) {
-                sensor.inputs.push_back(i.lockFunctor());
-            }
-        }
-        return result;
-    }
+    virtual cbox::CboxError streamFrom(cbox::DataIn& in) override final;
 
-    void
-    writeMessage(blox_TempSensorCombi& message, bool includeReadOnly) const
-    {
-        FieldTags stripped;
+    virtual cbox::CboxError streamTo(cbox::DataOut& out) const override final;
 
-        message.sensors_count = sensor.inputs.size();
-        message.combinefunc = _blox_SensorCombiFunc(sensor.func);
-        for (uint8_t i = 0; i < message.sensors_count && i < 8; i++) {
-            message.sensors[i] = inputs[i].getId();
-        }
+    virtual cbox::CboxError streamPersistedTo(cbox::DataOut& out) const override final;
 
-        if (includeReadOnly) {
-            if (sensor.valid()) {
-                message.value = cnl::unwrap((sensor.value()));
-            } else {
-                stripped.add(blox_TempSensorCombi_value_tag);
-            }
-        }
-        stripped.copyToMessage(message.strippedFields, message.strippedFields_count, 1);
-    }
+    virtual cbox::update_t update(const cbox::update_t& now) override final;
 
-    virtual cbox::CboxError
-    streamTo(cbox::DataOut& out) const override final
-    {
-        blox_TempSensorCombi message = blox_TempSensorCombi_init_zero;
-        writeMessage(message, true);
+    virtual void* implements(const cbox::obj_type_t& iface) override final;
 
-        return streamProtoTo(out, &message, blox_TempSensorCombi_fields, blox_TempSensorCombi_size);
-    }
-
-    virtual cbox::CboxError
-    streamPersistedTo(cbox::DataOut& out) const override final
-    {
-        blox_TempSensorCombi message = blox_TempSensorCombi_init_zero;
-        writeMessage(message, false);
-        return streamProtoTo(out, &message, blox_TempSensorCombi_fields, blox_TempSensorCombi_size);
-    }
-
-    virtual cbox::update_t
-    update(const cbox::update_t& now) override final
-    {
-        sensor.update();
-        return update_1s(now);
-    }
-
-    virtual void*
-    implements(const cbox::obj_type_t& iface) override final
-    {
-        if (iface == BrewBloxTypes_BlockType_TempSensorCombi) {
-            return this; // me!
-        }
-        if (iface == cbox::interfaceId<TempSensor>()) {
-            // return the member that implements the interface in this case
-            TempSensor* ptr = &sensor;
-            return ptr;
-        }
-        return nullptr;
-    }
-
-    TempSensorCombi&
-    get()
+    TempSensorCombi& get()
     {
         return sensor;
     }
+
+private:
+    void writeMessage(blox_TempSensorCombi& message, bool includeReadOnly) const;
 };
